requirements.txt
discord-py-interactions
discord.py[voice]
